requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
dotenv==0.9.9
ecdsa==0.19.1
Faker==37.5.3
fastapi==0.116.1
h11==0.16.0
idna==3.10
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
# Editable install with no version control (volleyball-team-creator==0.1.0)
-e /Users/skleinhans/Desktop/Test-Lab/volleyball-team-creator
